fix: dashboard widgets
