+++ src/assetbundles/src/js/RecentEntries.js
@@ -82,20 +82,20 @@
 
                             this.$widget.find('#recent-entries-widget').html(widgetHtml);
                         }
-    
+
                         window.translationsdashboard.widgets[widgetId].updateContainerHeight();
                         window.translationsdashboard.grid.refreshCols(true, true);
-    
+
                         $('.new-entry-check .checkbox').on('click', function(e) {
                             $(e.target).closest('tr[id^=item-entry-]').toggleClass('sel');
                             Craft.Translations.RecentEntries.prototype.updateSelected();
                         });
-    
+
                         $('.view-diff-entry').on('click', function(e) {
                             e.preventDefault();
-    
+
                             var diffHtml = content[$(e.target).attr('data-id')].diff;
-    
+
                             var classNames = [
                                 'entryId',
                                 'entryName',
@@ -104,22 +104,22 @@
                                 'entryDate',
                                 'wordDifference'
                             ];
-    
+
                             // Show the modal
                             $modal_entry.show();
-    
+
                             // Set modification details
                             for (let index = 0; index < classNames.length; index++) {
                                 $('.' + classNames[index]).html(content[$(e.target).attr('data-id')][classNames[index]]);
                             }
-    
+
                             // Add the diff html
                             document.getElementById("modal-body-entry").innerHTML = diffHtml;
-    
+
                             $('#modal-body-entry').on('click', 'div.diff-copy', function(event) {
                                 Craft.Translations.OrderEntries.copyTextToClipboard(event);
                             });
-    
+
                             $('#close-diff-modal-entry').on('click', function(e) {
                                 e.preventDefault();
                                 $modal_entry.hide();
