+++ src/services/App.php
@@ -115,6 +115,11 @@ class App extends Component
      * @var repository\UserRepository
      */
     public $userRepository;
+    
+    /**
+     * @var repository\WidgetRepository
+     */
+    public $widgetRepository;
 
     /**
      * @var repository\StaticTranslationsRepository
@@ -188,6 +193,7 @@ class App extends Component
         $this->orderRepository = new repository\OrderRepository();
         $this->translatorRepository = new repository\TranslatorRepository();
         $this->userRepository = new repository\UserRepository();
+        $this->widgetRepository = new repository\WidgetRepository();
         $this->staticTranslationsRepository = new repository\StaticTranslationsRepository();
         $this->wordCounter = new WordCounter();
         $this->fieldTranslatorFactory = new fieldtranslator\Factory();
