+++ src/templates/_index.twig
@@ -27,7 +27,7 @@
                 <ul>
                     {% for type, info in widgetTypes|filter(info => info.selectable) %}
                         <li>
-                            <a data-type="{{ type }}" 
+                            <a data-type="{{ type }}"
                                 data-name="{{ info.name }}">
                                 <span class="icon">{{ info.iconSvg|raw }}</span>
                                 {{ info.name == 'Modified Source Entries' ? 'New & Modified Entries' : info.name }}
