+++ src/widgets/Translators.php
@@ -33,7 +33,7 @@ class Translators extends Widget
     /**
      * @inheritdoc
      */
-    public static function iconPath()
+    public static function icon(): ?string
     {
         return Craft::getAlias('@acclaro/translations/icon-mask.svg');
     }
@@ -44,7 +44,7 @@ class Translators extends Widget
     public function rules(): array
     {
         $rules = parent::rules();
-        
+
         return $rules;
     }
 
@@ -62,9 +62,9 @@ class Translators extends Widget
     public function getBodyHtml(): ?string
     {
         $view = Craft::$app->getView();
-        
+
         $translators = $this->_getTranslators();
-        
+
         return $view->renderTemplate('translations/_components/widgets/Translators/body', ['translators' => $translators]);
     }
